Move button to sep folder
