+++ front/app/containers/Authentication/steps/_components/AuthProviderButton/Consent.tsx
@@ -9,7 +9,7 @@ import Error from 'components/UI/Error';
 import { useIntl, FormattedMessage } from 'utils/cl-intl';
 import Link from 'utils/cl-router/Link';
 
-import messages from './messages';
+import messages from '../../AuthProviders/messages';
 
 const Container = styled.div`
   display: flex;
