+++ front/app/containers/Authentication/steps/_components/AuthProviderButton/tracks.ts
@@ -1,0 +1,4 @@
+export default {
+  signInWithSSOClicked: 'Sign in with SSO button clicked',
+  signUpWithSSOClicked: 'Sign up with SSO button clicked',
+};
