+++ front/app/modules/commercial/id_vienna_saml/components/ViennaSamlButton.tsx
@@ -4,7 +4,7 @@ import { fontSizes, Box, Text } from '@citizenlab/cl2-component-library';
 import styled from 'styled-components';
 
 import { StyledAuthProviderButton } from 'containers/Authentication/steps/AuthProviders';
-import { TOnContinueFunction } from 'containers/Authentication/steps/AuthProviders/AuthProviderButton';
+import { TOnContinueFunction } from 'containers/Authentication/steps/_components/AuthProviderButton';
 import { SignUpInFlow } from 'containers/Authentication/typings';
 
 import { FormattedMessage } from 'utils/cl-intl';
